requirements.txt
django==1.10.4
django-auth-ldap==1.2.8
djangorestframework==3.5.3
djangorestframework-jsonapi==2.1.1
djangorestframework-jwt==1.9.0
django-filter==1.0.1
django-crispy-forms==1.6.1
django-jet==1.0.4
psycopg2
pytz
